+++ client/index.html
@@ -4,9 +4,10 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React + TS</title>
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css" rel="stylesheet">
+    <title>Badminton Academy</title>
   </head>
-  <body class="p-5 h-screen text-center bg-gray-900 text-white relative">
+  <body class="text-center bg-gray-900 text-white">
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
