+++ client/src/App.tsx
@@ -1,15 +1,17 @@
 import './App.css'
-import SignUp from './components/SignUp'
-import LogIn from './components/Login'
-import Navbar from './components/Navbar'
+import Home from './components/Home';
+import SignUp from './components/auth/SignUp'
+import LogIn from './components/auth/Login'
+import Navbar from './components/navbar/Navbar'
 import { Route, Routes } from 'react-router-dom'
 
 
 function App() {
   return (
-    <div className='flex flex-col justify-center items-center gap-10'>
+    <div className='h-screen flex items-center justify-center'>
       <Navbar/>
       <Routes>
+      <Route path='/' element= {<Home/>}></Route>
         <Route path='/Signup' element= {<SignUp/>}></Route>
         <Route path='/Login' element= {<LogIn/>}></Route>
       </Routes>
