+++ client/src/components/Home.tsx
@@ -1,7 +1,11 @@
+import Carousel from "./sections/carasol";
+import About from "./sections/About";
 
 function LandingPage(){
     return(
         <>
+            <Carousel/>
+            <About/>
         </>
     )
 }
